pyproject.toml
[build-system]
requires = [
    "setuptools>=60",
    "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rv_profile"
version = "0.2.0"
description = "A tool for generating flamegraphs of RISC-V programs running on RTL simulations"
readme = "README.md"
authors = [{ name = "Vincenzo Petrolo", email = "vincenzo.petrolo@polito.it" }]
license = { text = "MIT" }
dependencies = [
    "pylibfst==0.2.0",
    "wal-lang==0.8.2"
]
requires-python = ">=3.7"

[project.scripts]
rv_profile = "rv_profile.cli:main"  # CLI entry point

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["*.pl"]